+++ src/main/kotlin/com/github/pool_party/pull_party_bot/commands/handlers/AliasCommand.kt
@@ -9,6 +9,7 @@ import com.github.pool_party.pull_party_bot.commands.messages.ON_ALIAS_PARSE_FAI
 import com.github.pool_party.pull_party_bot.commands.messages.ON_CREATE_REQUEST_FAIL
 import com.github.pool_party.pull_party_bot.commands.messages.ON_PARTY_NAME_FAIL
 import com.github.pool_party.pull_party_bot.commands.messages.onAliasSuccess
+import com.github.pool_party.pull_party_bot.commands.sendMessageLogging
 import com.github.pool_party.pull_party_bot.database.dao.AliasCreationResult
 import com.github.pool_party.pull_party_bot.database.dao.ChatDao
 import com.github.pool_party.pull_party_bot.database.dao.PartyDao
@@ -16,32 +17,31 @@ import com.github.pool_party.pull_party_bot.database.dao.PartyDao
 class AliasCommand(
     private val partyDao: PartyDao,
     chatDao: ChatDao,
-) : CaseCommand("alias", "create a party alias", HELP_ALIAS, chatDao) {
+) : CaseCommand("alias", "create a new party with the same users", HELP_ALIAS, chatDao) {
 
     override suspend fun Bot.action(message: Message, args: String?) {
         val parsedArgs = parseArgs(args)
         val chatId = message.chat.id
 
         if (parsedArgs.isNullOrEmpty() || parsedArgs.size != 2) {
-            sendMessage(chatId, ON_ALIAS_PARSE_FAIL, "Markdown")
+            sendMessageLogging(chatId, ON_ALIAS_PARSE_FAIL)
             return
         }
 
         val (aliasName, partyName) = parsedArgs
 
         if (!validatePartyName(aliasName)) {
-            sendMessage(chatId, ON_PARTY_NAME_FAIL, "Markdown")
+            sendMessageLogging(chatId, ON_PARTY_NAME_FAIL)
             return
         }
 
-        sendMessage(
+        sendMessageLogging(
             chatId,
             when (partyDao.createAlias(chatId, aliasName, partyName)) {
                 AliasCreationResult.SUCCESS -> onAliasSuccess(aliasName)
                 AliasCreationResult.NAME_TAKEN -> ON_CREATE_REQUEST_FAIL
                 AliasCreationResult.NO_PARTY -> ON_ALIAS_FAIL
             },
-            "Markdown"
         )
     }
 }