+++ src/main/kotlin/com/github/pool_party/pull_party_bot/commands/handlers/DeleteCommands.kt
@@ -14,20 +14,21 @@ import com.github.pool_party.pull_party_bot.commands.messages.ON_DELETE_EMPTY
 import com.github.pool_party.pull_party_bot.commands.messages.onAliasDeleteSuccess
 import com.github.pool_party.pull_party_bot.commands.messages.onPartyDeleteSuggest
 import com.github.pool_party.pull_party_bot.commands.messages.onPartyDeleteUnchanged
+import com.github.pool_party.pull_party_bot.commands.sendMessageLogging
 import com.github.pool_party.pull_party_bot.database.dao.ChatDao
 import com.github.pool_party.pull_party_bot.database.dao.PartyDao
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 
 class DeleteCommand(private val partyDao: PartyDao, chatDao: ChatDao) :
-    AdministratorCommand("delete", "forget the parties as they have never happened", HELP_DELETE, chatDao) {
+    AdministratorCommand("delete", "delete the parties you provided", HELP_DELETE, chatDao) {
 
     override fun Bot.mainAction(message: Message, args: String?) {
         val parsedArgs = parseArgs(args)?.distinct()
         val chatId = message.chat.id
 
         if (parsedArgs.isNullOrEmpty()) {
-            sendMessage(chatId, ON_DELETE_EMPTY, "Markdown")
+            sendMessageLogging(chatId, ON_DELETE_EMPTY)
             return
         }
 
@@ -81,6 +82,6 @@ class ClearCommand(chatDao: ChatDao) :
     override fun Bot.mainAction(message: Message, args: String?) {
         val chatId = message.chat.id
         chatDao.clear(chatId)
-        sendMessage(chatId, ON_CLEAR_SUCCESS, "Markdown")
+        sendMessageLogging(chatId, ON_CLEAR_SUCCESS)
     }
 }