+++ src/main/kotlin/com/github/pool_party/pull_party_bot/commands/handlers/RudeCommand.kt
@@ -6,6 +6,7 @@ import com.github.pool_party.pull_party_bot.commands.CaseCommand
 import com.github.pool_party.pull_party_bot.commands.messages.HELP_RUDE
 import com.github.pool_party.pull_party_bot.commands.messages.ON_RUDE_FAIL
 import com.github.pool_party.pull_party_bot.commands.messages.onRudeSuccess
+import com.github.pool_party.pull_party_bot.commands.sendMessageLogging
 import com.github.pool_party.pull_party_bot.database.dao.ChatDao
 
 class RudeCommand(chatDao: ChatDao) : CaseCommand("rude", "switch RUDE(CAPS LOCK) mode", HELP_RUDE, chatDao) {
@@ -18,7 +19,7 @@ class RudeCommand(chatDao: ChatDao) : CaseCommand("rude", "switch RUDE(CAPS LOCK
             "on" -> chatDao.setRude(chatId, true)
             "off" -> chatDao.setRude(chatId, false)
             else -> {
-                sendMessage(chatId, ON_RUDE_FAIL, "Markdown")
+                sendMessageLogging(chatId, ON_RUDE_FAIL)
                 return
             }
         }