+++ src/main/kotlin/com/github/pool_party/pull_party_bot/database/Models.kt
@@ -17,7 +17,7 @@ object Aliases : IntIdTable() {
     val chatId = reference("chat_id", Chats, ReferenceOption.CASCADE, ReferenceOption.CASCADE)
 
     val name = varchar("name", 50)
-    val lastUse = datetime("last_use").defaultExpression(CurrentDateTime())
+    val lastUse = datetime("last_use").defaultExpression(CurrentDateTime)
 
     init {
         index(false, name, chatId)
@@ -25,7 +25,6 @@ object Aliases : IntIdTable() {
     }
 }
 
-object Chats : LongIdTable() {
-    override val id = long("chat_id").entityId()
+object Chats : LongIdTable(columnName = "chat_id") {
     val isRude = bool("is_rude").default(false)
 }