+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.2.2] - 18.04.2022
+
+### Fixed
+
+- Using logging everywhere
+- Escaping markdown in messages
+
 ## [1.2.1] - 04.04.2022
 
 ### Added
@@ -110,6 +117,7 @@
 -   Implicit party pulling
 -   Rude mode
 
+[1.2.2]: https://github.com/pool-party/pull-party-bot/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/pool-party/pull-party-bot/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/pool-party/pull-party-bot/compare/v1.1.0.2...v1.2.0
 [1.1.0.2]: https://github.com/pool-party/pull-party-bot/compare/v1.1.0.1...v1.1.0.2